Reorganized shared lib files; adjusted paths; moved shared utilities from individual apps
