+++ apps/ui/src/app/services/auth.service.ts
@@ -1,11 +1,11 @@
 import { HttpClient } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
-import { AuthKeys } from '@gh/shared';
+import { AuthKeys } from '@gh/shared/models';
+import { loggedMethod } from '@gh/shared/utils';
 import { CookieService } from 'ngx-cookie-service';
 import { catchError, of, tap } from 'rxjs';
 import { StoreService } from 'services/store.service';
 import { publicPost, refreshPost } from 'utils/api';
-import { loggedMethod } from 'utils/decorators';
 
 type Credentials = {
 	accessToken: string;
@@ -50,7 +50,7 @@ export class AuthService {
 		return this.credentials.refreshToken;
 	}
 
-	@loggedMethod
+	@loggedMethod()
 	login(email: string, password: string) {
 		const url = `${this.#baseApiUrl}/auth/login`;
 		const credentials = { email, password };
@@ -71,20 +71,20 @@ export class AuthService {
 			);
 	}
 
-	@loggedMethod
+	@loggedMethod()
 	loginGoogle() {
 		const url = `${this.#baseApiUrl}/auth/google/login`;
 
 		window.location.href = url;
 	}
 
-	@loggedMethod
+	@loggedMethod()
 	logout() {
 		this.clearCredentials();
 		this.#storeService.resetUserCards();
 	}
 
-	@loggedMethod
+	@loggedMethod()
 	refresh(refreshToken: string) {
 		const url = `${this.#baseApiUrl}/auth/refresh`;
