+++ apps/ui/src/app/utils/index.ts
@@ -1,4 +1,2 @@
 export * from './api';
-export * from './decorators';
 export * from './dom';
-
