+++ libs/shared/src/lib/index.ts
@@ -1,0 +1,1 @@
+export * from './models';
